+++ spec/rails_app/app/controllers/sessions_controller.rb
@@ -8,4 +8,12 @@ class SessionsController < ApplicationController
     session[:admin] = false
     redirect_to root_url, :notice => "Successfully signed out"
   end
+
+  def count
+    session[:count] ||= 0
+    if request.post?
+      session[:count] += 1
+    end
+    render :layout => false
+  end
 end