+++ spec/rails_app/config/routes.rb
@@ -12,6 +12,8 @@ RailsApp::Application.routes.draw do
   namespace :sessions do
     get 'signin'
     get 'signout'
+    get 'count'
+    post 'count'
   end
   # The priority is based upon order of creation:
   # first created -> highest priority.